+++ packages/pluggableWidgets/badge-native/src/__tests__/Badge.spec.tsx
@@ -4,10 +4,12 @@ import { fireEvent, render } from "@testing-library/react-native";
 
 import { Badge, Props } from "../Badge";
 
-jest.mock("react-native/Libraries/Utilities/Platform", () => ({
-    OS: "ios",
-    select: jest.fn(dict => dict.ios)
-}));
+jest.mock("react-native/Libraries/Utilities/Platform", () => {
+    const Platform = jest.requireActual("react-native/Libraries/Utilities/Platform");
+    Platform.OS = "ios";
+    Platform.default = { ...Platform.default, OS: "ios" };
+    return Platform;
+});
 
 const defaultProps: Props = {
     name: "badge-test",